[Fleet] Implement agent integration health reporting
